Only regenerate dependencies if necessary

+++ calculate-dependency-size.sh
@@ -1,10 +1,13 @@
 #!/bin/bash
 
-echo stage2/01-sys-tweaks/00-packages \
-  | xargs cat \
-  | tr ' ' '\n' \
-  | xargs -L 1 debtree \
-  | tee dependency_graph.txt
+if [ ! -f dependency_graph.txt ]; then
+  # echo stage2/01-sys-tweaks/00-packages \
+  find . -type f -iname "*packages*" \
+    | xargs cat \
+    | tr ' ' '\n' \
+    | xargs -L 1 debtree \
+    | tee dependency_graph.txt
+fi
 
 cat dependency_graph.txt \
   | grep -o -E '"\w+" -> "\w+"' \
@@ -12,7 +15,7 @@ cat dependency_graph.txt \
   | sort \
   | uniq \
   | xargs apt-cache show \
-  | grep -E "^(Package|Size)" \
+  | grep -E "^(Package|Installed-Size)" \
   | cut -d " " -f 2 \
   | xargs -L 2 echo \
   | sed -E -e 's/(.*) (.*)/\2 \1/g' \
